Create program24.py
Given a number N find whether its a happy number or not. A number is called happy if it leads to 1 after a sequence of steps wherein each step, the number is replaced by the sum of squares of its digit that is if we start with Happy Number and keep replacing it with digits square sum, we reach 1. 

Example 1: 
Input: N = 19 
Output: 1 
Explanation: 19 is a Happy Number, 1^2 + 9^2 = 82 8^2 + 2^2 = 68 6^2 + 8^2 = 100 1^2 + 0^2 + 0^2 = 1 As we reached to 1, 19 is a Happy Number. 

Example 2: 
Input: N = 20 
Output: 0 
Explanation: We can never reach to 1 by repeating the mentioned steps.