Create program25.py
Ravi is an enthusiastic student. He knows that in computer terminology what a palindrome means. We can have palindromes with numbers as well as strings. For example, madam is a string palindrome and 121 is a number palindrome. Ravi has been assigned a task by his teacher. In this task, he has been given a numbers. His duty is to tell the given number is palindrome or Not.If Palindrome print 'YES' otherwise Print 'NO'

 Sample Test Case 

Test case 1: 
Input 121 
Output YES